@@ -11,14 +11,16 @@
     "prepublish": "npm run build",
     "lint": "eslint src",
     "pretest": "npm run lint",
-    "test": "node --require dotenv/config node_modules/.bin/ava"
+    "test": "node --require dotenv/config node_modules/.bin/ava",
+    "fix:prettier": "prettier \"src/**/*.js\" --write",
+    "fix": "run-s fix:*"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/mshick/librabbitmq.git"
   },
   "engines": {
-    "node": ">=6.0.0"
+    "node": ">=8.0.0"
   },
   "keywords": [
     "rabbitmq",
@@ -41,6 +43,10 @@
       "babel-register"
     ]
   },
+  "prettier": {
+    "singleQuote": true,
+    "proseWrap": "always"
+  },
   "dependencies": {
     "amqplib": "^0.5.2",
     "lodash": "^4.4.2",
@@ -49,18 +55,22 @@
     "uuid": "^3.0.1"
   },
   "devDependencies": {
-    "ava": "^0.24.0",
+    "ava": "^0.25.0",
     "babel-cli": "^6.24.1",
     "babel-eslint": "^8.1.2",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.4.0",
     "babel-preset-stage-1": "^6.24.1",
     "babel-register": "^6.24.1",
-    "dotenv": "^4.0.0",
-    "eslint": "^4.3.0",
-    "eslint-config-parsley": "^1.2.1",
-    "eslint-plugin-filenames": "^1.2.0",
-    "eslint-plugin-import": "^2.7.0",
+    "dotenv": "^6.0.0",
+    "eslint": "^5.1.0",
+    "eslint-config-parsley": "^1.2.2",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-plugin-filenames": "^1.3.2",
+    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-prettier": "^2.6.2",
+    "npm-run-all": "^4.1.3",
+    "prettier": "^1.13.7",
     "rimraf": "^2.6.1"
   }
 }
