@@ -18,10 +18,10 @@ const defaultSubscriberOptions = {
   exclusive: true
 };
 
-const addSubscriber = async function (args, plugin) {
-  const {connection, exchange, topic, subscriber, options} = args;
-  const {state: pluginState} = plugin;
-  const {_openChannels} = pluginState;
+const addSubscriber = async function(args, plugin) {
+  const { connection, exchange, topic, subscriber, options } = args;
+  const { state: pluginState } = plugin;
+  const { _openChannels } = pluginState;
 
   const {
     channelName: userChannelName,
@@ -29,27 +29,45 @@ const addSubscriber = async function (args, plugin) {
     exchangeOptions: userExchangeOptions,
     queueOptions: userQueueOptions,
     subscriberOptions: userSubscriberOptions
-  } = options || {};
+  } =
+    options || {};
 
   let channel;
 
   try {
-    const channelName = userChannelName || getChannelName({method: 'subscribeToMessages', exchange});
+    const channelName =
+      userChannelName ||
+      getChannelName({ method: 'subscribeToMessages', exchange });
     const exchangeTopic = topic ? `${exchange}.${topic}` : `${exchange}.*`;
 
-    const exchangeOptions = defaultsDeep({}, userExchangeOptions, defaultExchangeOptions);
-    const queueOptions = defaultsDeep({}, userQueueOptions, defaultQueueOptions);
-    const subscriberOptions = defaultsDeep({}, userSubscriberOptions, defaultSubscriberOptions);
+    const exchangeOptions = defaultsDeep(
+      {},
+      userExchangeOptions,
+      defaultExchangeOptions
+    );
+    const queueOptions = defaultsDeep(
+      {},
+      userQueueOptions,
+      defaultQueueOptions
+    );
+    const subscriberOptions = defaultsDeep(
+      {},
+      userSubscriberOptions,
+      defaultSubscriberOptions
+    );
 
     if (_openChannels[channelName]) {
       channel = _openChannels[channelName].channel;
     } else {
-      channel = await createChannel({
-        name: channelName,
-        options: channelOptions,
-        persist: true,
-        connection
-      }, plugin);
+      channel = await createChannel(
+        {
+          name: channelName,
+          options: channelOptions,
+          persist: true,
+          connection
+        },
+        plugin
+      );
     }
 
     await channel.assertExchange(exchange, 'topic', exchangeOptions);
@@ -65,7 +83,11 @@ const addSubscriber = async function (args, plugin) {
       options: subscriberOptions
     });
 
-    const consumed = await channel.consume(queue.queue, consumer, subscriberOptions);
+    const consumed = await channel.consume(
+      queue.queue,
+      consumer,
+      subscriberOptions
+    );
 
     return {
       channel,
